+++ lib/downloader.js
@@ -26,33 +26,6 @@ function safeMkdir(dir) {
     }
 }
 
-class BuiltinModule {
-    constructor(id, manifest, deviceClass) {
-        this._id = id;
-        this._manifest = manifest;
-        this._loaded = deviceClass;
-        deviceClass.metadata = manifest;
-    }
-
-    get id() {
-        return this._id;
-    }
-    get manifest() {
-        return this._manifest;
-    }
-    get version() {
-        return 0; // version does not matter for builtin
-    }
-
-    clearCache() {
-        // nothing to do here
-    }
-
-    getDeviceFactory() {
-        return this._loaded;
-    }
-}
-
 module.exports = class ModuleDownloader {
     constructor(platform, client, builtins = {}) {
         this._platform = platform;
@@ -215,7 +188,7 @@ module.exports = class ModuleDownloader {
             if (manifest.module_type === 'org.thingpedia.builtin') {
                 if (!this._builtins[id])
                     throw new Error(`The device ${id} is not supported in this instance of ThingSystem`);
-                return new BuiltinModule(id, manifest, this._builtins[id].module);
+                return new Modules['org.thingpedia.builtin'](id, manifest, this, this._builtins[id].module);
             }
 
             console.log('Loaded manifest for ' + id + ', module type: '+ manifest.module_type + ', version: ' + manifest.version);
