+++ lib/modules/base_js.js
@@ -2,7 +2,7 @@
 //
 // This file is part of ThingEngine
 //
-// Copyright 2017 The Board of Trustees of the Leland Stanford Junior University
+// Copyright 2017-2019 The Board of Trustees of the Leland Stanford Junior University
 //
 // Author: Giovanni Campagna <gcampagn@cs.stanford.edu>
 //
@@ -10,45 +10,14 @@
 "use strict";
 
 const assert = require('assert');
-const path = require('path');
-const fs = require('fs');
-const tmp = require('tmp');
 const stream = require('stream');
-
-const Module = require('module');
 
 const Tp = require('thingpedia');
 
 const { makeGenericOAuth } = require('./utils');
 const { ImplementationError } = require('./errors');
 
-// shared code between all modules with custom JS
-
-function resolve(mainModule) {
-    if (!mainModule.startsWith('/'))
-        throw new Error('Invalid relative module path');
-    if (require.resolve)
-        return require.resolve(mainModule);
-    else
-        return Module._resolveFilename(mainModule, module, false);
-}
-
-function clearRequireCache(mainModule) {
-    try {
-        var fileName = resolve(mainModule);
-        console.log(mainModule + ' was cached as ' + fileName);
-
-        delete require.cache[fileName];
-
-        var prefix = path.dirname(fileName) + '/';
-        for (var key in require.cache) {
-            if (key.startsWith(prefix))
-                delete require.cache[key];
-        }
-    } catch(e) {
-        // do nothing
-    }
-}
+// shared code between all modules with JS implementation
 
 function isIterable(result) {
     return typeof result === 'object' && result !== null &&
@@ -99,8 +68,6 @@ module.exports = class BaseJavascriptModule {
 
         this._loader = loader;
         this._client = loader.client;
-        this._platform = loader.platform;
-        this._cacheDir = loader.platform.getCacheDir() + '/device-classes';
         this._id = id;
         this._manifest = manifest;
 
@@ -117,26 +84,34 @@ module.exports = class BaseJavascriptModule {
     get version() {
         return this._manifest.version;
     }
-    get package_version() {
-        return this._manifest.package_version;
+
+    static get [Symbol.species]() {
+        // TRICKY NOTE: in JS, "static" means class method, not static method
+        // so we can use "this", and "this" refers to the class where this accessor
+        // is invoked; hence, if this accessor is called on a derived class, we
+        // will return the derived class, which is the desired behavior (and also
+        // how this feature is used by native Array and native RegExp)
+        return this;
     }
 
+    /* istanbul ignore next */
     clearCache() {
-        this._loading = null;
-
-        if (this._modulePath)
-            clearRequireCache(this._modulePath);
+        throw new TypeError('Not Implemented');
     }
 
-    _createSubmodule(id, manifest, deviceClass) {
+    /* istanbul ignore next */
+    _doGetDeviceFactory() {
+        throw new TypeError('Not Implemented');
+    }
+
+    async _createSubmodule(id, manifest, deviceClass) {
         const submodule = new (this.constructor[Symbol.species])(id, manifest, this._loader);
-        submodule._loading = deviceClass;
-        submodule._completeLoading(deviceClass);
+        await submodule._completeLoading(deviceClass);
 
         return submodule;
     }
 
-    _completeLoading(deviceClass) {
+    async _completeLoading(deviceClass) {
         deviceClass.metadata = this._manifest;
 
         if (this._manifest.auth.type === 'oauth2' && !deviceClass.runOAuth2)
@@ -145,18 +120,18 @@ module.exports = class BaseJavascriptModule {
             deviceClass.runOAuth2.install(deviceClass.prototype);
 
         for (let action in this.manifest.actions) {
-            if (typeof module.prototype['do_' + action] !== 'function')
+            if (typeof deviceClass.prototype['do_' + action] !== 'function')
                 throw new ImplementationError(`Implementation for action ${action} missing`);
-            deviceClass.prototype['do_' + action] = safeWrapAction(module.prototype['do_' + action]);
+            deviceClass.prototype['do_' + action] = safeWrapAction(deviceClass.prototype['do_' + action]);
         }
         for (let query in this.manifest.queries) {
             const pollInterval = this.manifest.queries[query].poll_interval;
-            if (pollInterval === 0 && typeof module.prototype['subscribe_' + query] !== 'function')
+            if (pollInterval === 0 && typeof deviceClass.prototype['subscribe_' + query] !== 'function')
                 throw new ImplementationError(`Poll interval === 0 but no subscribe function was found`);
             if (typeof deviceClass.prototype['get_' + query] !== 'function')
                 throw new ImplementationError(`Implementation for query ${query} missing`);
 
-            deviceClass.prototype['get_' + query] = safeWrapQuery(module.prototype['get_' + query], query);
+            deviceClass.prototype['get_' + query] = safeWrapQuery(deviceClass.prototype['get_' + query], query);
             if (!deviceClass.prototype['subscribe_' + query]) {
                 if (pollInterval > 0) {
                     deviceClass.prototype['subscribe_' + query] = function(params, state, filter) {
@@ -168,7 +143,7 @@ module.exports = class BaseJavascriptModule {
                     };
                 }
             } else {
-                deviceClass.prototype['subscribe_' + query] = safeWrapSubscribe(module.prototype['subscribe_' + query], query);
+                deviceClass.prototype['subscribe_' + query] = safeWrapSubscribe(deviceClass.prototype['subscribe_' + query], query);
             }
             if (!deviceClass.prototype['history_' + query]) {
                 deviceClass.prototype['history_' + query] = function(params, base, delta, filters) {
@@ -181,96 +156,26 @@ module.exports = class BaseJavascriptModule {
                 };
             }
         }
-    }
-
-    _loadJsModule(id) {
-        var modulePath = this._modulePath;
-        var version = JSON.parse(fs.readFileSync(modulePath + '/package.json').toString('utf8'))['thingpedia-version'];
-        if (version !== this._manifest.package_version) {
-            console.log(`Cached module ${this.id} is out of date (found ${version}, want ${this._manifest.package_version})`);
-            return null;
-        }
-
-        var deviceClass = require(modulePath);
-        deviceClass.require = function(subpath) {
-            return require(path.resolve(modulePath, subpath));
-        };
-        try {
-            this._completeLoading(deviceClass);
-        } catch(e) {
-            return Promise.reject(e);
-        }
 
         const subdevices = deviceClass.subdevices || {};
-        return Promise.all((this.manifest.child_types || []).map((childId) => {
+        await Promise.all((this.manifest.child_types || []).map(async (childId) => {
             if (!(childId in subdevices)) {
                 console.error(`Child device ${childId} is not declared in ${deviceClass.name}.subdevices, this will cause unexpected behavior`);
-                return Promise.resolve();
+                return;
             }
 
-            return this._loader.loadManifest(childId, true).then((childManifest) => {
-                const submodule = this._createSubmodule(childId, childManifest, subdevices[childId]);
-                this._loader.injectModule(childId, submodule);
-            });
-        })).then(() => {
-            this._loading = deviceClass;
-            return deviceClass;
-        });
-    }
+            const childManifest = await this._loader.loadManifest(childId, true);
+            const submodule = await this._createSubmodule(childId, childManifest, subdevices[childId]);
+            this._loader.injectModule(childId, submodule);
+        }));
 
-    async _doGetDeviceFactory() {
-        try {
-            const redirect = await this._client.getModuleLocation(this._id, this._manifest.version);
-            const response = await Tp.Helpers.Http.getStream(redirect);
-
-            const [path, fd] = await new Promise((resolve, reject) => {
-                tmp.file({ mode: 0o600,
-                           keep: true,
-                           dir: this._platform.getTmpDir(),
-                           prefix: 'thingengine-' + this._id + '-',
-                           postfix: '.zip' }, (err, path, fd, cleanup) => {
-                    if (err)
-                        reject(err);
-                    else
-                        resolve([path, fd, cleanup]);
-                });
-            });
-            const stream = fs.createWriteStream('', { fd, flags: 'w' });
-            const zipPath = await new Promise((callback, errback) => {
-                response.pipe(stream);
-                stream.on('finish', () => {
-                    callback(path);
-                });
-                stream.on('error', errback);
-            });
-            try {
-                fs.mkdirSync(this._modulePath);
-            } catch(e) {
-                if (e.code !== 'EEXIST')
-                    throw e;
-            }
-
-            var unzip = this._platform.getCapability('code-download');
-            await unzip.unzip(zipPath, this._modulePath);
-            fs.unlinkSync(zipPath);
-            return await this._loadJsModule();
-        } catch (e) {
-            this._loading = null;
-            throw e;
-        }
+        this._loading = deviceClass;
+        return deviceClass;
     }
 
     getDeviceFactory() {
         if (this._loading)
             return Promise.resolve(this._loading);
-
-        this._modulePath = path.resolve(process.cwd(), this._cacheDir + '/' + this._id);
-
-        if (fs.existsSync(this._modulePath)) {
-            var cached = this._loadJsModule();
-            if (cached)
-                return Promise.resolve(this._loading = cached);
-        }
 
         return this._loading = this._doGetDeviceFactory();
     }
