+++ lib/modules/index.js
@@ -13,8 +13,11 @@ const V2 = require('./v2');
 
 const GenericRestModule = require('./generic');
 const RSSModule = require('./rss_factory');
+const BuiltinModule = require('./builtin');
 
 module.exports = {
+    'org.thingpedia.builtin': BuiltinModule,
+
     'org.thingpedia.v2': V2,
     'org.thingpedia.rss': RSSModule,
     'org.thingpedia.generic_rest.v1': GenericRestModule
