+++ lib/modules/v2.js
@@ -9,10 +9,7 @@
 // See LICENSE for details
 "use strict";
 
-const BaseJavascriptModule = require('./base_js');
+const BaseOnDiskJavascriptModule = require('./base_ondisk_js');
 
-module.exports = class ThingpediaModuleV2 extends BaseJavascriptModule {
-    static get [Symbol.species]() {
-        return ThingpediaModuleV2;
-    }
+module.exports = class ThingpediaModuleV2 extends BaseOnDiskJavascriptModule {
 };
